Split resolver validation pass
